Add `install` script
This script helps installing hugo and the theme locally. It works for
Linux and macOS, both 64bit. If the system already has an extended
`hugo` binary present in the `PATH`, it will be use it, and skip
installing it again.
It also installs the PyBCN theme from GH, without requiring git.
This will be useful for both local and build environments.